+++ CHANGES.rst
@@ -1,6 +1,14 @@
 CHANGE LOG
 ==========
 
+5.6.4
+-----
+- Bug Fix: add yourwords check to surveys list
+
+5.6.3
+-----
+- Bug Fix: removed yourwords surveys from template tag lists
+
 5.6.2
 -----
 - Bug Fix: remove PreventDeleteMixin from Ts&Cs index page